+++ src/commands/neuron_manage.rs
@@ -9,6 +9,7 @@ use candid::{CandidType, Encode, Principal};
 use clap::{Parser, ValueEnum};
 use ic_base_types::PrincipalId;
 use ic_nns_common::pb::v1::{NeuronId, ProposalId};
+use ic_nns_governance::pb::v1::manage_neuron::RefreshVotingPower;
 use ic_nns_governance::pb::v1::{
     manage_neuron::{
         configure::Operation, disburse::Amount, AddHotKey, ChangeAutoStakeMaturity, Command,
@@ -138,6 +139,11 @@ pub struct ManageOpts {
     /// a limited subset (Private).
     #[arg(long)]
     set_visibility: Option<NativeVisibility>,
+
+    /// Refresh the neuron's voting power by reaffirming the current list of followers.
+    /// This must be done every so often to avoid neurons diminishing in voting power.
+    #[arg(long)]
+    refresh_followers: bool,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, ValueEnum)]
@@ -152,7 +158,7 @@ pub fn exec(auth: &AuthInfo, opts: ManageOpts) -> AnyhowResult<Vec<IngressWithRe
             opts.add_hot_key.is_none() && opts.remove_hot_key.is_none() && !opts.disburse && opts.disburse_amount.is_none() && opts.disburse_to.is_none()
             && !opts.clear_manage_neuron_followees && !opts.join_community_fund && !opts.leave_community_fund
             && opts.follow_topic.is_none() && opts.follow_neurons.is_none() && opts.register_vote.is_none() && !opts.reject
-            && opts.set_visibility.is_none(),
+            && opts.set_visibility.is_none() && !opts.refresh_followers,
             "\
 Cannot use --ledger with these flags. This version of quill only supports the following neuron-manage operations with a Ledger device:
 --additional-dissolve-delay-seconds, --start-dissolving, --stop-dissolving, --split, --merge-from-neuron, --spawn, --stake-maturity, --auto-stake-maturity"
@@ -413,6 +419,15 @@ Cannot use --ledger with these flags. This version of quill only supports the fo
             id: None,
         })?;
 
+        msgs.push(args);
+    }
+
+    if opts.refresh_followers {
+        let args = Encode!(&ManageNeuron {
+            command: Some(Command::RefreshVotingPower(RefreshVotingPower {})),
+            neuron_id_or_subaccount: id.clone(),
+            id: None,
+        })?;
         msgs.push(args);
     }